pyproject.toml
[tool.poetry]
name = "bot_aiogram"
version = "0.1.0"
description = ""
authors = ["Sergey Vavilin <vavilin.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.13.1"
python-decouple = "^3.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
